fix: simplify init methods and add docs example
