+++ packages/hooks/src/utils/queryParams.ts
@@ -1,7 +1,6 @@
-import { isArray, isEmpty, isNumber, isSSR } from '@sajari/react-sdk-utils';
+import { isArray, isNumber, isSSR } from '@sajari/react-sdk-utils';
 
-import { combineFilters, FilterBuilder, Range, RangeFilterBuilder, Variables } from '../ContextProvider/controllers';
-import { ProviderPipelineState, SyncURLState } from '../ContextProvider/types';
+import { FilterBuilder, Range, RangeFilterBuilder, Variables } from '../ContextProvider/controllers';
 
 export function isRange(value: unknown) {
   return isArray(value) && value.length === 2 && isNumber(value[0]) && isNumber(value[1]);
@@ -26,74 +25,50 @@ export function getSearchParams() {
   return search;
 }
 
-export const initParam = ({
+export const initFiltersFromURLState = ({
   filters,
-  variables,
-  searchState,
-  autocompleteState,
-  defaultFilter,
-  syncURLState,
+  params,
 }: {
   filters: (FilterBuilder | RangeFilterBuilder)[];
-  variables: Variables;
-  searchState: ProviderPipelineState;
-  autocompleteState: ProviderPipelineState;
-  defaultFilter?: string;
-  syncURLState?: SyncURLState;
+  params: Record<string, string>;
 }) => {
-  const params = getSearchParams();
-
-  if (syncURLState) {
-    filters.forEach((filter) => {
-      if (filter instanceof FilterBuilder) {
-        const key = filter.getField() || filter.getName();
-        const value = params[key] || '';
-        filter.set(value ? value.split(',') : []);
-      } else if (filter instanceof RangeFilterBuilder) {
-        const key = filter.getField() || filter.getName();
-        const value = params[key] || '';
-        const initialRange = paramToRange(value);
-        const limit = (params[`${key}_min_max`] || '').split(':').map(Number) as Range;
-        if (isRange(initialRange)) {
-          filter.set(initialRange as Range);
-        }
-        if (isRange(limit)) {
-          filter.setMin(limit[0]);
-          filter.setMax(limit[1]);
-          // Freeze the state of the filterBuilder to avoid the UI from being overridden at the first response
-          filter.setFrozen(true);
-        }
+  filters.forEach((filter) => {
+    if (filter instanceof FilterBuilder) {
+      const key = filter.getField() || filter.getName();
+      const value = params[key] || '';
+      filter.set(value ? value.split(',') : []);
+    } else if (filter instanceof RangeFilterBuilder) {
+      const key = filter.getField() || filter.getName();
+      const value = params[key] || '';
+      const initialRange = paramToRange(value);
+      const limit = (params[`${key}_min_max`] || '').split(':').map(Number) as Range;
+      if (isRange(initialRange)) {
+        filter.set(initialRange as Range);
       }
-    });
-  }
-
-  const filter = combineFilters(filters ?? []);
-  const variablesFilterString = variables.get().filter ?? '';
-  const defaultFilterString = defaultFilter?.toString() ?? '';
-
-  variables.set({
-    filter: () => {
-      const expression = filter.filter();
-      return [defaultFilterString, variablesFilterString, isEmpty(expression) ? '_id != ""' : expression]
-        .filter(Boolean)
-        .join(' AND ');
-    },
-    countFilters: () => filter.countFilters(),
-    buckets: () => filter.buckets(),
-    count: () => filter.count(),
-    min: () => filter.min(),
-    max: () => filter.max(),
-    ...(syncURLState
-      ? {
-          ...(params.show ? { resultsPerPage: params.show } : {}),
-          ...(params.sort ? { sort: params.sort } : {}),
-          ...(params[autocompleteState.config.qParam]
-            ? { [autocompleteState.config.qParam]: params[autocompleteState.config.qParam] }
-            : {}),
-          ...(params[searchState.config.qParam]
-            ? { [searchState.config.qParam]: params[searchState.config.qParam] }
-            : {}),
-        }
-      : {}),
+      if (isRange(limit)) {
+        filter.setMin(limit[0]);
+        filter.setMax(limit[1]);
+        // Freeze the state of the filterBuilder to avoid the UI from being overridden at the first response
+        filter.setFrozen(true);
+      }
+    }
   });
+};
+
+export const initVariableFromURLState = ({
+  variables,
+  mappingKeys,
+  params,
+}: {
+  variables: Variables;
+  mappingKeys: { paramKey: string; variableKey: string; defaultValue?: string }[];
+  params: Record<string, string>;
+}) => {
+  const variableObj = {};
+  mappingKeys.forEach(({ paramKey, variableKey, defaultValue }) => {
+    if (params[paramKey]) {
+      variableObj[variableKey] = params[paramKey] || defaultValue;
+    }
+  });
+  variables.set(variableObj);
 };
